sismo: fix clippy lints across sismo-core and sismo-cli
Apply clippy's idiomatic-Rust suggestions: drop needless same-type casts and
`.into()`s, remove now-unnecessary `unsafe` around the safe `__cpuid_count`
intrinsic, use `writeln!`/`div_ceil`/`is_multiple_of`, replace a
`p.is_empty()` guard with a `Some("")` pattern, iterate with `enumerate`
instead of a hand-rolled counter, and group the hex literal digits evenly.

A couple were fallout from earlier work: the `_ =>` data-source arm made
KIND_SISMO_VENDOR lib-unused (inlined its wire value in the one test), and the
libc signal-handler casts now go through a fn-pointer type so clippy's
fn_to_numeric_cast is satisfied. `intern_mapping` keeps its 8 args behind an
explicit allow, and the session-lock open states `.truncate(false)` (it
truncates after flock, not before).

Left for a follow-up: the `# Safety` docs on the vestigial `extern "C"`
helpers (no C++ caller remains, so those should become plain Rust fns) and
some macOS-only dead-code warnings in the datasource command.

Validated on Linux (sismo-core 63 + sismo-cli 15 tests, a record with the
privileged marker) and cross-checked for aarch64-apple-darwin with cargo-zigbuild.

+++ crates/sismo-core/src/symbolize/disasm.rs
@@ -164,13 +164,13 @@ fn decode_x86(bytes: &[u8], func_start: u64, cb: InsnCb, ctx: *mut c_void) {
     let decoder = yaxpeax_x86::amd64::InstDecoder::default();
     let mut reader = U8Reader::new(bytes);
     loop {
-        let before = u64::from(<U8Reader as Reader<u64, u8>>::total_offset(&mut reader)) as usize;
+        let before = <U8Reader as Reader<u64, u8>>::total_offset(&mut reader) as usize;
         if before >= bytes.len() {
             break;
         }
         match decoder.decode(&mut reader) {
             Ok(inst) => {
-                let after = u64::from(<U8Reader as Reader<u64, u8>>::total_offset(&mut reader)) as usize;
+                let after = <U8Reader as Reader<u64, u8>>::total_offset(&mut reader) as usize;
                 if after <= before {
                     break;
                 }
@@ -186,13 +186,13 @@ fn decode_arm(bytes: &[u8], func_start: u64, cb: InsnCb, ctx: *mut c_void) {
     let decoder = yaxpeax_arm::armv8::a64::InstDecoder::default();
     let mut reader = U8Reader::new(bytes);
     loop {
-        let before = u64::from(<U8Reader as Reader<u64, yaxpeax_arch::U32le>>::total_offset(&mut reader)) as usize;
+        let before = <U8Reader as Reader<u64, yaxpeax_arch::U32le>>::total_offset(&mut reader) as usize;
         if before >= bytes.len() {
             break;
         }
         match decoder.decode(&mut reader) {
             Ok(inst) => {
-                let after = u64::from(<U8Reader as Reader<u64, yaxpeax_arch::U32le>>::total_offset(&mut reader)) as usize;
+                let after = <U8Reader as Reader<u64, yaxpeax_arch::U32le>>::total_offset(&mut reader) as usize;
                 if after <= before {
                     break;
                 }
